src/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fonttools==4.61.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.2.6
openai==2.24.0
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
python-dotenv==0.21.0
